feat: tabs파일의 내용을 k-culture의 index 내용으로 변경 후 기존 k-culture 파일 삭제
